+++ code/src/cljs/sixsq/nuvla/ui/clouds/events.cljs
@@ -1,11 +1,11 @@
-(ns sixsq.nuvla.ui.infrastructures.events
+(ns sixsq.nuvla.ui.clouds.events
   (:require
     [cljs.spec.alpha :as s]
     [re-frame.core :refer [dispatch reg-event-db reg-event-fx]]
     [sixsq.nuvla.ui.cimi-api.effects :as cimi-api-fx]
     [sixsq.nuvla.ui.cimi-detail.events :as cimi-detail-events]
-    [sixsq.nuvla.ui.infrastructures.spec :as spec]
-    [sixsq.nuvla.ui.infrastructures.utils :as utils]
+    [sixsq.nuvla.ui.clouds.spec :as spec]
+    [sixsq.nuvla.ui.clouds.utils :as utils]
     [sixsq.nuvla.ui.main.spec :as main-spec]
     [sixsq.nuvla.ui.messages.events :as messages-events]
     [sixsq.nuvla.ui.utils.general :as general-utils]