+++ code/src/cljs/sixsq/nuvla/ui/dashboard/utils.cljs
@@ -17,7 +17,7 @@
 (def target-navigator {:resource        "apps"
                        :tab-index       apps-store-utils/tab-navigator
                        :tab-index-event ::apps-store-events/set-active-tab-index})
-(def target-nbs {:resource "edge"})
+(def target-nbs {:resource "edges"})
 (def target-creds {:resource "credentials"})
 
 (defn type->icon