+++ code/src/cljs/sixsq/nuvla/ui/welcome/views.cljs
@@ -134,7 +134,7 @@
               (@tr [:welcome-how-to-nb-2-subheader-pre])
               " "
               ^{:key (@tr [:welcome-how-to-nb-2-subheader-mid])}
-              [navigate-link "edge" (@tr [:welcome-how-to-nb-2-subheader-mid])]
+              [navigate-link "edges" (@tr [:welcome-how-to-nb-2-subheader-mid])]
               " "
               (@tr [:welcome-how-to-nb-2-subheader-post])
               ".")
@@ -154,7 +154,7 @@
               (@tr [:welcome-how-to-nb-4-subheader-pre])
               " "
               ^{:key (@tr [:welcome-how-to-nb-4-subheader-mid])}
-              [navigate-link "edge" (@tr [:welcome-how-to-nb-4-subheader-mid])]
+              [navigate-link "edges" (@tr [:welcome-how-to-nb-4-subheader-mid])]
               ". "
               (@tr [:welcome-how-to-nb-4-subheader-post])
               ".")
@@ -241,7 +241,7 @@
               (@tr [:welcome-how-to-video-2-subheader-pre])
               " "
               ^{:key (@tr [:welcome-how-to-video-2-subheader-post])}
-              [navigate-link "edge" (@tr [:welcome-how-to-video-2-subheader-post])] ".")
+              [navigate-link "edges" (@tr [:welcome-how-to-video-2-subheader-post])] ".")
         (step 3
               (@tr [:welcome-how-to-video-3-header])
               (@tr [:welcome-how-to-video-3-subheader-pre]))]]]]))